Add method to create route with matrix

+++ Models/ShortestRoute.cs
@@ -6,9 +6,9 @@
         public List<string> pointsNames { get; set; }
         public double distance { get; set; }
         public double duration { get; set; }
-        public double totalPrice { get; set; }
+        public decimal totalPrice { get; set; }
 
-        public double[] priceDistribution { get; set; }
+        public decimal[] priceDistribution { get; set; }
 
         public ShortestRoute()
         {
